Refactor document attachment handling to DocumentAttachment concern
